% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenergyvol_fun.R, R/allpeakdemand_fun.R
\name{allenergyvol_fun}
\alias{allenergyvol_fun}
\title{Energy volume functions}
\usage{
allenergyvol_fun(dnsp.in, other.df, pdprof.df, projyearend.in,
  energyvol.df, use.in)

allenergyvol_fun(dnsp.in, other.df, pdprof.df, projyearend.in,
  energyvol.df, use.in)
}
\arguments{
\item{dnsp.in}{DNSP selected for function loop}

\item{other.df}{database of static dnsp inputs}

\item{pdprof.df}{database of static dnsp inputs for peak demand profile}

\item{projyearend.in}{dynamic variable of final year}

\item{energyvol.df}{database of energy vol function ouput}

\item{use.in}{dynamic variable of growth of energy use per customer}

\item{cars.in}{dynamic variable for electric vehicle penetration by 2060 as a percent}

\item{cust.in}{dynamic variable of growth of customers percent}

\item{ogrid.in}{dynamic variable of offgrid customers by 2060 percent}

\item{solar.in}{dynamic variable of percent of solar penetration by 2060}

\item{batt.in}{dynamic variable of percent of battery penetration by 2060}

\item{dnsp.in}{DNSP selected for function loop}

\item{other.df}{database of static dnsp inputs}

\item{projyearend.in}{dynamic variable of final year}

\item{use.in}{dynamic variable of growth of energy use per customer}
}
\description{
Formula for calculating the energy volume distribution among customers

Formula for calculating the energy volume distribution among customers
}
