% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgrowth_fun.R
\name{allgrowth_fun}
\alias{allgrowth_fun}
\title{Growth capex model functions}
\usage{
allgrowth_fun(dnsp.in, other.df, pdprof.df, projyearend.in, energyvol.df,
  peakdemand.df, demand.in, solar.in, batt.in, age.in, cust.in, cars.in)
}
\arguments{
\item{dnsp.in}{DNSP selected for function loop}

\item{other.df}{database of static dnsp inputs}

\item{pdprof.df}{database of static dnsp inputs for peak demand profile}

\item{projyearend.in}{dynamic variable of final year}

\item{energyvol.df}{database of energy volumes function ouput}

\item{peakdemand.df}{database of peak demand function ouput}

\item{demand.in}{dynamic variable of demand management intensity}

\item{solar.in}{dynamic variable of percent customers with solar by 2060}

\item{batt.in}{dynamic variable of percent customers with batteries by 2060}

\item{age.in}{dynamic variable of age of assets}

\item{cust.in}{dynamic variable of customer numbers growth}

\item{cars.in}{dynamic variable of proportion of electric vehicle penetration by 2060}
}
\description{
Formula for calculating the growth capex model
}
