% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allpeakdemand_fun.R
\name{allpeakdemand_fun}
\alias{allpeakdemand_fun}
\title{Peak demand functions}
\usage{
allpeakdemand_fun(dnsp.in, other.df, pdprof.df, projyearend.in,
  energyvol.df, use.in, signal.in)
}
\arguments{
\item{dnsp.in}{DNSP selected for function loop}

\item{other.df}{database of static dnsp inputs}

\item{pdprof.df}{database of static dnsp inputs for peak demand profile}

\item{projyearend.in}{dynamic variable of final year}

\item{energyvol.df}{database of energy vol function ouput}

\item{use.in}{dynamic variable of growth of energy use per customer}

\item{signal.in}{dynamic variable of dynamic price signal}
}
\description{
Formula for calculating the peak demand profile with dynamic price signal
}
