% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carenergyvol_fun.R
\name{carenergyvol_fun}
\alias{carenergyvol_fun}
\title{Car penetraion function}
\usage{
carenergyvol_fun(yearslabel, other.df, electriccars)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{other.df}{database of static dnsp inputs}

\item{electriccars}{number of electric vehicles}
}
\description{
Formula for calculating the penetration of electric vehicles over time
}
