% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carevolchange_fun.R
\name{carevolchange_fun}
\alias{carevolchange_fun}
\title{Change in electric vehicle energy volume function}
\usage{
carevolchange_fun(yearslabel, carenergyvol)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{carenergyvol}{vector of absolute energy volumes from electric vehicles from carenergyvol_fun}
}
\description{
Formula for calculating the change in electric vehicle energy volume
}
