% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars_fun.R
\name{cars_fun}
\alias{cars_fun}
\title{Car numbers function}
\usage{
cars_fun(yearslabel, other.df)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{other.df}{database of static dnsp inputs}
}
\description{
Formula for calculating the cars in each dnsp region
}
