% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif.R
\name{cif_fun}
\alias{cif_fun}
\title{CIF function}
\usage{
cif_fun(y, infl, startyearend, projyearend)
}
\arguments{
\item{y}{cifyear vector of 1:noyears+1 cif is moved across one year}

\item{infl}{inflation from dnsp other.df dataframe}

\item{startyearend}{always 2019 or 19}

\item{projyearend}{dynamic input to define the end year of the modelling}
}
\description{
Cumulative inflation risk for each year in model
}
\details{
Prerequisites that need to be defined in ptrm model (i.e. run above code):
cifyear, infl, startyearend, projyearend
}
\examples{
projyearend = 97
startyearend = 19
noyears = projyearend-startyearend
infl=0.02425
cifyears=1:noyears+1


}
\keyword{cif,}
\keyword{inflation}
