% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons_nonsolar.R
\name{cons_nonsolar_fun}
\alias{cons_nonsolar_fun}
\title{Consumption load profile for non-solar customers}
\usage{
cons_nonsolar_fun(yearslabel, priceprof.df, unit.rate, signal.in)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{priceprof.df}{dataframe of static load profile for year 1}

\item{unit.rate}{energy unit rate $/kwh}

\item{signal.in}{dynamic variable of dynamic price signal}
}
\description{
Formula for calculating the consumption load profile for non-solar customers
#'
}
