% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iab.R
\name{iab_fun}
\alias{iab_fun}
\title{Initial asset base function}
\usage{
iab_fun(y, remlife, oavalue, assetclass)
}
\arguments{
\item{y}{vector of years from 1}

\item{remlife}{remaining life of asset from assets.df}

\item{oavalue}{opening asset value from assets.df}

\item{assetclass}{asset class number}
}
\description{
Formula for initial asset base for a single asset. All assets must be calculated by looping this function
}
\details{
This is used in the year on year depreciation calculations.
PTRM excel has iab as a macro for each asset class e.g. asset class 1 is A1iab:G76.
Returns the iab for one asset class for all years.

Prerequisites that need to be defined in ptrm model (i.e. run above code):
remlife,oavalue,assetclass
}
\examples{

projyearend = 97
startyearend = 19
noyears = projyearend-startyearend
y = seq(1:noyears)
assetclass = 1
oavalue = 1812.17
remlife = 30.70

}
\keyword{asset}
\keyword{capex,}
\keyword{class,}
\keyword{iab,}
\keyword{oavalue,}
\keyword{on}
\keyword{remlife,}
\keyword{tracking,}
\keyword{year}
