% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadafteff_fun.R
\name{loadafteff_fun}
\alias{loadafteff_fun}
\title{Load after energy efficiency function}
\usage{
loadafteff_fun(yearslabel, other.df, use.in, loadaft_offgrid)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{other.df}{dataset of DNSP static vectors}

\item{use.in}{dynamic variable of energy use per customer growth rate}

\item{loadaft_offgrid}{energy load ater offgrid customers exited}
}
\description{
Formula for calculating the load of electricity after energy efficiencies
}
