% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopall_fun.R
\name{loopall_fun}
\alias{loopall_fun}
\title{Energy volume functions}
\usage{
loopall_fun(assets.df, other.df, iab.df, priceprof.df, projyearend.in,
  age.in, retireslim.in, addnew.in, productivity.in, dnsp.in, rba.in,
  use.in, cust.in, cars.in, ogrid.in, solar.in, batt.in, signal.in)
}
\arguments{
\item{assets.df}{database of static dnsp inputs for all assets}

\item{other.df}{database of static dnsp inputs}

\item{iab.df}{database of static dnsp inputs for iab and tax iab}

\item{priceprof.df}{dataframe of static dnsp price profile inputs}

\item{projyearend.in}{dynamic variable of final year}

\item{age.in}{dynamic variable of age of assets}

\item{retireslim.in}{dynamic variable of slim assets}

\item{addnew.in}{dynamic variable of add new assets}

\item{productivity.in}{dynamic variable of productivity}

\item{dnsp.in}{DNSP selected for function loop}

\item{rba.in}{dynamic variable of RBA cash rate}

\item{use.in}{dynamic variable of growth of energy use per customer}

\item{cust.in}{dynamic variable of growth of customers percent}

\item{cars.in}{dynamic variable for electric vehicle penetration by 2060 as a percent}

\item{ogrid.in}{dynamic variable of offgrid customers by 2060 percent}

\item{solar.in}{dynamic variable of percent of solar penetration by 2060}

\item{batt.in}{dynamic variable of percent of battery penetration by 2060}

\item{signal.in}{dynamic variable of price signal text input}
}
\description{
Formula for calculating the energy volume distribution among customers
}
