% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncagg.R
\name{ncagg_fun}
\alias{ncagg_fun}
\title{Netcapex functions: vi Aggregate}
\usage{
ncagg_fun(noassets, yearslabel, productivity, fcnetcapex, ncreplace.df,
  ncskinny.df, ncaugex.df, ncsyscap.df, ncnonnetwk.df)
}
\arguments{
\item{noassets}{count of asset classes}

\item{yearslabel}{label of years from 2020}

\item{productivity}{productivity DNSP yearly vector}

\item{fcnetcapex}{dataframe of PTRM input net capex for 2020-24}

\item{ncreplace.df}{dataframe of netcapex i replacement from ncreplace_fun}

\item{ncskinny.df}{dataframe of netcapex ii skinny from ncskinny_fun}

\item{ncaugex.df}{dataframe of netcapex iii augex from ncaugex_fun}

\item{ncsyscap.df}{dataframe of netcapex iv system capital overhead from ncsyscap_fun}

\item{ncnonnetwk.df}{dataframe of netcapex v non-network from ncnonnetwk_fun}
}
\description{
Final aggregate formula for extending netcapex beyond five years March 22 2019
}
\details{
Returns a database of projected netcapex:vi aggregate for all assets across all years.
Prerequisites that need to be defined in ptrm model (i.e. run above code) are defined in 'arguments' below.
}
