% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncreplaceqld.R
\name{ncreplaceqld_fun}
\alias{ncreplaceqld_fun}
\title{Netcapex functions: i Replacement}
\usage{
ncreplaceqld_fun(yearslabel, fcnetavg, noyears, projyearend, noassets,
  assetcode, repratio, age, modrepcost, startincline, ratedecline)
}
\arguments{
\item{yearslabel}{label pf projected years}

\item{fcnetavg}{empty dataframe with first 5 years from DNSP and an avg for 2025}

\item{noyears}{count of years}

\item{projyearend}{final projected year}

\item{noassets}{count of asset classes}

\item{assetcode}{if asset type is coded as 1 or 2}

\item{repratio}{replacement ratio percent, DNSP Static variable}

\item{age}{age of assets in years, DNSP yearly vector}

\item{modrepcost}{modern day replacement cost, DNSP Static variable}

\item{startincline}{Start incline of net capex, DNSP Static variable}

\item{ratedecline}{Rate of decline, DNSP Static variable}
}
\description{
First formula for extending netcapex beyond five years March 22 2019
}
\details{
Returns a database of projected netcapex: i replacement for all assets across all years.

Prerequisites that need to be defined in ptrm model (i.e. run above code) are defined in arguments below.
}
