% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncskinny.R
\name{ncskinny_fun}
\alias{ncskinny_fun}
\title{Netcapex functions: ii Skinny}
\usage{
ncskinny_fun(noassets, yearslabel, fcnetcapex, ncreplace.df, retireslim)
}
\arguments{
\item{noassets}{count of asset classes}

\item{yearslabel}{label of years from 2020}

\item{fcnetcapex}{dataframe of PTRM input net capex for 2020 to 2024}

\item{ncreplace.df}{dataframe of replacement net capex calculate from function ncreplace_fun}

\item{retireslim}{retire or slim assets as a percent of prev year, DNSP yearly vector}
}
\description{
Second formula for extending netcapex beyond five years March 22 2019
}
\details{
Returns a database of projected netcapex:ii skinny for all assets across all years.

Prerequisites that need to be defined in ptrm model (i.e. run above code) are defined in arguments below.
}
