% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncsyscap.R
\name{ncsyscap_fun}
\alias{ncsyscap_fun}
\title{Netcapex functions: iv System capital overhead}
\usage{
ncsyscap_fun(noassets, yearslabel, fcnetavg, syscapratio, assetcode)
}
\arguments{
\item{noassets}{count of asset classes}

\item{yearslabel}{label of years from 2020}

\item{fcnetavg}{blank dataframe of PTRM input net capex for 2020 to end with PTRM inputs}

\item{syscapratio}{system capital overhead ratio, DNSP static variable}

\item{assetcode}{if asset class is coded as 1 or 2}
}
\description{
Fourth formula for extending netcapex beyond five years March 22 2019
}
\details{
Returns a database of projected netcapex:iv system capital overhead for all assets across all years.

Prerequisites that need to be defined in ptrm model (i.e. run above code) are defined in arguments below.
}
