% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcapex.R
\name{netcapex_fun}
\alias{netcapex_fun}
\title{New netcapex function}
\usage{
netcapex_fun(fcnetcapex, yearslabel, noassets, noyears, age, productivity,
  networksize)
}
\arguments{
\item{fcnetcapex}{dataframe of first 5 years of netcapex for dnsp}

\item{yearslabel}{from 2020 to projected year end}

\item{noassets}{count of asset classes}

\item{noyears}{count of years}

\item{age}{dynamic input age of assets}

\item{productivity}{dynamic input productivity}

\item{networksize}{dynamic input network size}
}
\description{
Formula for extending netcapex beyond five years
}
\details{
Returns a database of projected netcapex for all assets x all years.
2025 netcapex is first determined by averaging the years from 2020-24.
For the following years a dnsp-specific formula is used to predict netcapex until projected year end.

Prerequisites that need to be defined in ptrm model (i.e. run above code):
fcnetcapex:2020-2024 dataframe, yearslabel, noassets, noyears
and the dynamic inputs: age, productivity and networksize. Currently only defined for Ausgrid.
Eventually this will by updated to specify a unique formula for each asset class type and dnsp.
}
\examples{

projyearend = 97
startyearend = 19
noyears = projyearend-startyearend
noassets = 22
age = 50
productivity = 0
networksize = 0

}
\keyword{asset}
\keyword{capex,}
\keyword{class}
\keyword{class,}
\keyword{dnsp,}
\keyword{netcapex,}
\keyword{type}
