% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettax.R
\name{nettax_fun}
\alias{nettax_fun}
\title{Net tax functions}
\usage{
nettax_fun(yearslabel, startyearend, projyearend, other.df, tax.df)
}
\arguments{
\item{yearslabel}{label of years from year two to projected year end}

\item{startyearend}{start of projections}

\item{projyearend}{final projected year}

\item{other.df}{dataframe of PTRM input other values}

\item{tax.df}{dataframe of tax calculations for 2019 onwards}
}
\description{
Formulas for calculating the net tax allowance by years
}
