% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcust_fun.R
\name{newcust_fun}
\alias{newcust_fun}
\title{Customer numbers function}
\usage{
newcust_fun(yearslabel, custnum)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{custnum}{vector of customer numbers from cust_fun}
}
\description{
Formula for calculating the number of customers in each dnsp
}
