% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opex.R
\name{opex_fun}
\alias{opex_fun}
\title{OPEX function}
\usage{
opex_fun(yearslabel, controlopex, productivity, other.df)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{controlopex}{vector of controllable opex for 2020-24 from DNSP ptrm input}

\item{productivity}{productivity percent, DNSP dynamic yearly vector}

\item{other.df}{dataset of DNSP static vectors}
}
\description{
Formula for calculating the controllable OPEX for all projected years
}
\details{
Returns a vector of OPEX for each year.

Prerequisites that need to be defined in ptrm model (i.e. run above code) are in arguments below:
}
