% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opexdrc.R
\name{opexdrc_fun}
\alias{opexdrc_fun}
\title{OPEX function}
\usage{
opexdrc_fun(yearslabel, opexdrc)
}
\arguments{
\item{yearslabel}{label of projection years from year start}

\item{opexdrc}{vector of opex debt raising cost for first 5 years from DNSP ptrm input}
}
\description{
Formula for calculating the controllable OPEX for all projected years
}
\details{
Returns a vector of OPEX for each year.

Prerequisites that need to be defined in ptrm model (i.e. run above code) are in arguments below:
}
