% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbatteries_fun.R
\name{pbatt_fun}
\alias{pbatt_fun}
\title{Proportion of customers with batteries}
\usage{
pbatt_fun(other.df, yearslabel, batt.in)
}
\arguments{
\item{other.df}{DNSP other.df dataframe}

\item{yearslabel}{label of projection years from 2020}

\item{batt.in}{dynamic input of prop customers offgrid by 2060}
}
\description{
Proportion of customers with batters in each year
}
\keyword{customers}
\keyword{solar,}
