% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdbattery.R
\name{pdbattery_fun}
\alias{pdbattery_fun}
\title{Peak demand profile for batteries}
\usage{
pdbattery_fun(yearslabel, pdprof.df, energyvol.df, pdsolar)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{pdprof.df}{peak demand profile per hour input dataframe}

\item{energyvol.df}{modeled outputs of energy volumes from allenergyvol_fun}

\item{pdsolar}{peak demand adjusted for offgrid, ee and solar}
}
\description{
Formula for calculating the peak demand profile after adjusting for off grid customers, energy efficiency,
solar and batteries
}
