% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdenergyeffic.R
\name{pdenergyeffic_fun}
\alias{pdenergyeffic_fun}
\title{Peak demand profile for energy efficiency}
\usage{
pdenergyeffic_fun(yearslabel, use.in, pdoffgrid)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{use.in}{dynamic variable of energy use}

\item{pdoffgrid}{peak demand adjusted for off grid customers}
}
\description{
Formula for calculating the peak demand profile for energy efficiency
}
