% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdevehicles.R
\name{pdevehicles_fun}
\alias{pdevehicles_fun}
\title{Peak demand profile for electric vehicles}
\usage{
pdevehicles_fun(yearslabel, pdprof.df, energyvol.df, pdbatt)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{pdprof.df}{peak demand profile per hour input dataframe}

\item{energyvol.df}{modeled outputs of energy volumes from allenergyvol_fun}

\item{pdbatt}{peak demand adjusted for offgrid, ee, solar and batteries}
}
\description{
Formula for calculating the peak demand profile after adjusting for off grid customers, energy efficiency,
solar, batteries and electric vehicles
}
