% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfinalsignal.R
\name{pdfinalsignal_fun}
\alias{pdfinalsignal_fun}
\title{Final dynamic price signal}
\usage{
pdfinalsignal_fun(yearslabel, pd_ev, signal.in)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{pd_ev}{peak demand adjusted for offgrid, ee, solar, batteries and ev}

\item{signal.in}{dynamic variable of dynamic price signal}
}
\description{
Formula for calculating the final dynamic price signal over hours and years
}
