% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdoffgrid.R
\name{pdoffgrid_fun}
\alias{pdoffgrid_fun}
\title{Peak demand profile for customers on grid}
\usage{
pdoffgrid_fun(yearslabel, energyvol.df, pdprof.df)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{energyvol.df}{dataset calculated from energy volumes function}

\item{pdprof.df}{dataset of DNSP static vectors}
}
\description{
Formula for calculating the peak demand profile for on grid customers
}
