% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poffgrid_fun.R
\name{poffgrid_fun}
\alias{poffgrid_fun}
\title{Proportion of offgrid customers function}
\usage{
poffgrid_fun(other.df, yearslabel, ogrid.in)
}
\arguments{
\item{other.df}{DNSP other.df dataframe}

\item{yearslabel}{label of projection years from 2020}

\item{ogrid.in}{dynamic input of prop customers offgrid by 2060}
}
\description{
Proportion of customers offgrid in each year
}
\keyword{offgrid}
