% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psolar_fun.R
\name{psolar_fun}
\alias{psolar_fun}
\title{Proportion of customers with solar}
\usage{
psolar_fun(other.df, yearslabel, solar.in)
}
\arguments{
\item{other.df}{DNSP other.df dataframe}

\item{yearslabel}{label of projection years from 2020}

\item{solar.in}{dynamic input of prop customers offgrid by 2060}
}
\description{
Proportion of customers with solar in each year
}
\keyword{customers}
\keyword{solar,}
