% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punitrate.R
\name{punitrate_fun}
\alias{punitrate_fun}
\title{Energy unit rate by hour}
\usage{
punitrate_fun(yearslabel, unit.rate, signal.in)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{unit.rate}{energy unit rate $/kwh}

\item{signal.in}{dynamic variable of dynamic price signal}
}
\description{
Formula for calculating the energy unit rate per hour for calculating price
}
