% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realcapex.R
\name{realcapex_fun}
\alias{realcapex_fun}
\title{Real net capex function}
\usage{
realcapex_fun(years, noassets, fcnetavg.full, rvanilla, yearslabel)
}
\arguments{
\item{years}{number of all years from 1}

\item{noassets}{count of asset classes}

\item{fcnetavg.full}{dataframe of all projected netcapex for dnsp (from netcapex_fun)}

\item{rvanilla}{real vanilla WACC vector for all years}

\item{yearslabel}{from 2020 to projected year end}
}
\description{
Formula for sum of real net capex in each year across all asset classes
}
\details{
Returns a vector of sumcapex for all years. Equal to excel PTRM assets row 10.
Sum of the half vanilla WACC of the netcapex of all assets in each year

Prerequisites that need to be defined in ptrm model (i.e. run above code):
fcnetavg.full: all years netcapex dataframe, years, noassets, fcnetavg.full, rvanilla, yearslabel
}
\examples{
projyearend = 97
startyearend = 19
years = seq(1:(projyearend-startyearend))
noassets = 22

}
\keyword{WACC}
\keyword{asset}
\keyword{capex,}
\keyword{class}
\keyword{class,}
\keyword{dnsp,}
\keyword{netcapex,}
\keyword{type,}
\keyword{vanilla}
