% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requity.R
\name{requity_fun}
\alias{requity_fun}
\title{Rate of equity function}
\usage{
requity_fun(other.df, yearslabel, rba)
}
\arguments{
\item{other.df}{DNSP other.df dataframe}

\item{yearslabel}{label of projection years from 2020}

\item{rba}{dynamic input of rba cash rate}
}
\description{
Varying rate of equity for each year
}
\details{
Rate of equity is static in ptrm by with new rba cash we need to vary it

Prerequisites that need to be defined in ptrm model (i.e. run above code):
year, rba, startyearend, projyearend
}
\keyword{debt}
\keyword{debt,}
\keyword{on}
\keyword{rba,}
\keyword{return}
