% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceloadbatt.R
\name{rloadbatt_fun}
\alias{rloadbatt_fun}
\title{Reduction in energy load from batteries}
\usage{
rloadbatt_fun(yearslabel, other.df, custongrid, pbatt)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{other.df}{dataset of DNSP static vectors}

\item{custongrid}{number of customers on grid}

\item{pbatt}{percent of customers with batteries}
}
\description{
Formula for calculating tbe cumulative reduction in energy load GWh from increased batteries per year
}
