% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceloadsolar.R
\name{rloadsolar_fun}
\alias{rloadsolar_fun}
\title{Reduction in energy load from solar}
\usage{
rloadsolar_fun(yearslabel, other.df, custongrid, psolar)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{other.df}{dataset of DNSP static vectors}

\item{custongrid}{number of customers on grid}

\item{psolar}{percent of customers on solar}
}
\description{
Formula for calculating tbe cumulative reduction in energy load GWh from increased solar per year
}
