% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRABend.R
\name{rRABend_fun}
\alias{rRABend_fun}
\title{Real RAB end period function}
\usage{
rRABend_fun(noyears, RABstart, deprec, realcapexall, startyearend,
  projyearend)
}
\arguments{
\item{noyears}{count of projection years}

\item{RABstart}{opening RAB PTRM input (oav+auc)}

\item{deprec}{sum of the regdeprec.all across all years for each asset class (WARL or yoy)}

\item{realcapexall}{vector of sum of capex over all asset categories for each year (realcapex_fun)}

\item{startyearend}{first year of model: 19 or 2019}

\item{projyearend}{end year of projection}
}
\description{
Formula for calculating the real residual RAB (end period)
}
\details{
Real residual RAB (end period) for each year. PTRM excel assets row 466.
real RAB = prevyearRAB - deprec + realcapex. Equal to the PTEM excel sheet assets row 466

Returns a vector of rRABend for each year.

Prerequisites that need to be defined in ptrm model (i.e. run above code):
noyears,RABstart,deprec,realcapexall
}
\keyword{RAB,}
\keyword{deprec,}
\keyword{real}
\keyword{realcapex}
