% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxiab.R
\name{taxiab_fun}
\alias{taxiab_fun}
\title{TAX: Initial asset base function}
\usage{
taxiab_fun(y, taxremlife, taxoavalue, assetclass)
}
\arguments{
\item{y}{vector of years from 1}

\item{taxremlife}{tax remaining life of asset from assets.df}

\item{taxoavalue}{tax opening asset value from assets.df}

\item{assetclass}{id of asset class to be modelled}
}
\description{
Formula for tax initial asset base for a single asset. All assets must be calculated by looping this function
}
\details{
This is used in the year on year tax depreciation calculations.
Returns the tax iab for one asset class for all years.
}
\keyword{asset}
\keyword{capex,}
\keyword{class,}
\keyword{iab,}
\keyword{oavalue,}
\keyword{on}
\keyword{remlife,}
\keyword{tax}
\keyword{tax,}
\keyword{tracking,}
\keyword{year}
