% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxyoydeprec.R
\name{taxyoy_fun}
\alias{taxyoy_fun}
\title{TAX: YOY depreciation function}
\usage{
taxyoy_fun(y, capexstage, assetclass, taxstdlife, startyearend, cif.df,
  fctaxcapex.df)
}
\arguments{
\item{y}{vector of years from 2019:projyearend, 2019 gets removed later}

\item{capexstage}{corresponds to the year number or 'years'}

\item{assetclass}{single asset class to model}

\item{taxstdlife}{standard life for the asset}

\item{startyearend}{2019 or 19}

\item{cif.df}{cumulative inflation index from 2020}

\item{fctaxcapex.df}{dataframe of the forecasted netcapex plus the customer contributions}
}
\description{
Additive formula for tax year on year depreciation in each year for a single asset.
All assets must be calculated by looping this function
}
\details{
This is used in the year on year depreciation calculations.
Returns the capex for one asset class for all years.

Prerequisites that need to be defined in ptrm model (i.e. run above code):
remlife,oavalue,assetclass
}
