% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoyall.R
\name{yoyall_fun}
\alias{yoyall_fun}
\title{WARL formulas}
\usage{
yoyall_fun(noassets, yearslabel, years, noyears, remlife, oavalue,
  startyearend, projyearend, stdlife, fcnetavg.full, rvanilla, iabcheck)
}
\arguments{
\item{noassets}{count of asset classes}

\item{yearslabel}{from 2020 to projected year end}

\item{years}{list of regulatory years from 1}

\item{noyears}{count of years}

\item{remlife}{vector of remaining life of assets from dnsp assets.df input dataframe}

\item{oavalue}{vector of opening asset value of assets from dnsp assets.df input dataframe}

\item{startyearend}{first regulatory year (2019 or 2020)}

\item{projyearend}{final projected year}

\item{stdlife}{vector of standard life of assets from dnsp assets.df input dataframe}

\item{fcnetavg.full}{dataframe of all netcapex for all assets defined by netcapex function}

\item{rvanilla}{vector of real vanilla WACC}

\item{iabcheck}{dataframe of raw iab values from ptrm for the first ten years}
}
\description{
Weighted average remaining life method for real depreciation
}
\details{
Function calculates only one asset class: ac, which is specified by user. The function has to be looped to project the WARL depreciation for all asset classes.
Prerequisites that need to be defined in ptrm model, i.e. run above code: ac, yearslabel, fcnetavg.full, rvanilla, remlife, stdlife, oavalue, noyears
}
\keyword{WACC}
\keyword{capex,}
\keyword{depreciation,}
\keyword{on}
\keyword{remlife,}
\keyword{stdlife,}
\keyword{vanilla}
\keyword{year}
\keyword{year,}
