% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoycapex.R
\name{yoycapex_fun}
\alias{yoycapex_fun}
\title{YOY capex function}
\usage{
yoycapex_fun(y, capexstage, noassets, oavalue, rvanilla, startyearend,
  stdlife, fcnetavg.full)
}
\arguments{
\item{y}{vector of years from 2019:projyearend, 2019 gets removed later}

\item{capexstage}{corresponds to the year number or 'years'}

\item{noassets}{count of asset classes}

\item{oavalue}{opening asset value from assets.df}

\item{rvanilla}{real vanilla WACC vector by years}

\item{startyearend}{2019 or 19}

\item{stdlife}{standard life for the asset}

\item{fcnetavg.full}{dataframe of the forecasted netcapex determined above}
}
\description{
Additive formula for capex depreciation in each year for a single asset.
All assets must be calculated by looping this function
}
\details{
This is used in the year on year depreciation calculations.
Returns the capex for one asset class for all years.

Prerequisites that need to be defined in ptrm model (i.e. run above code):
remlife,oavalue,assetclass
}
\examples{
projyearend = 97
startyearend = 19
y = (startyearend:projyearend)
stdlife = 22


}
\keyword{WACC}
\keyword{asset}
\keyword{capex,}
\keyword{class,}
\keyword{oavalue,}
\keyword{on}
\keyword{real}
\keyword{stdlife,}
\keyword{tracking,}
\keyword{vanilla}
\keyword{year}
