% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenergyvol_fun.R
\name{allenergyvol_fun}
\alias{allenergyvol_fun}
\title{Energy volume functions}
\usage{
allenergyvol_fun(dnsp.in, other.df, cars.in, projyearend.in)
}
\arguments{
\item{dnsp.in}{DNSP selected for function loop}

\item{other.df}{database of static dnsp inputs}

\item{cars.in}{dynamic variable for electric vehicle penetration by 2060 as a percent}

\item{projyearend.in}{dynamic variable of final year}
}
\description{
Formula for calculating the energy volume distribution among customers
}
