% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenergyvol_fun.R
\name{allenergyvol_fun}
\alias{allenergyvol_fun}
\title{Energy volume functions}
\usage{
allenergyvol_fun(yearslabel, other.df, electriccars, cars.in)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{other.df}{database of static dnsp inputs}

\item{electriccars}{number of electric vehicles}

\item{cars.in}{dynamic variable for electric vehicle penetration by 2060 as a percent}
}
\description{
Formula for calculating the energy volume distribution among customers
}
