% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cust_fun.R
\name{cust_fun}
\alias{cust_fun}
\title{Customer numbers function}
\usage{
cust_fun(yearslabel, other.df)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{other.df}{dataset of DNSP static vectors}
}
\description{
Formula for calculating the number of customers in each dnsp
}
