% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdsolar.R
\name{pdsolar_fun}
\alias{pdsolar_fun}
\title{Peak demand profile for solar}
\usage{
pdsolar_fun(yearslabel, pdprof.df, energyvol.df, pdeeffic)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{pdprof.df}{peak demand profile per hour input dataframe}

\item{energyvol.df}{modeled outputs of energy volumes from allenergyvol_fun}

\item{pdeeffic}{peak demand adjusted for energy efficiency}
}
\description{
Formula for calculating the peak demand profile after adjusting for off grid customers, energy efficiency and solar
}
