% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_fun.R
\name{price_fun}
\alias{price_fun}
\title{Price per GWh and per customer}
\usage{
price_fun(yearslabel, df.real, energyvol.df, other.df)
}
\arguments{
\item{yearslabel}{label of projection years from 2020}

\item{df.real}{dataframe of real revenue calculated from ptrm_fun}

\item{energyvol.df}{dataframe of relative energy volumes from allenergyvol_fun}

\item{other.df}{dataframe of static dnsp inputs}
}
\description{
Formula for calculating the relative price for customers
}
