% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_fun.R
\name{price_fun}
\alias{price_fun}
\title{Price per GWh and per customer}
\usage{
price_fun(dnsp.in, df.real, energyvol.df, other.df, projyearend.in,
  cust.in)
}
\arguments{
\item{dnsp.in}{DNSP selected for function loop}

\item{df.real}{dataframe of real revenue calculated from ptrm_fun}

\item{energyvol.df}{dataframe of relative energy volumes from allenergyvol_fun}

\item{other.df}{dataframe of static dnsp inputs}

\item{projyearend.in}{dynamic variable of final year}

\item{cust.in}{dynamic variable of customer growth percentage}
}
\description{
Formula for calculating the relative price for customers
}
