% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrm.R
\name{ptrm_fun}
\alias{ptrm_fun}
\title{PTRM all model}
\usage{
ptrm_fun(assets.df, other.df, projyearend.in, age.in, retireslim.in,
  addnew.in, productivity.in, dnsp.in, method, dynamicincdebt.in)
}
\arguments{
\item{assets.df}{dataframe of all asssets for all dnsps}

\item{other.df}{dataframe of all other inputs for all dnsps}

\item{projyearend.in}{dynamic input of final projected year}

\item{age.in}{dynamic input of asset age}

\item{retireslim.in}{dynamic input of retire or slim assets}

\item{addnew.in}{dynamic input of adding new assets}

\item{productivity.in}{dynamic input of productivity}

\item{dnsp.in}{dynamic input of selected dnsp}

\item{method}{dynamic input of method either yoy or warl}

\item{dynamicincdebt.in}{dynamic input of increase in debt}
}
\description{
Complete model with only the real output
}
\details{
Returns a database of projected revenue buildling blocks and capex for app

Prerequisites that need to be defined in ptrm model (i.e. run above code) are defined in arguments below.
}
\keyword{app}
\keyword{dynamic}
\keyword{model,}
\keyword{ptrm,}
\keyword{shiny,}
