% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdebt.R
\name{rdebt_fun}
\alias{rdebt_fun}
\title{Rate of debt function}
\usage{
rdebt_fun(other.df, yearslabel, dynamicincdebt)
}
\arguments{
\item{other.df}{DNSP other.df dataframe}

\item{yearslabel}{label of projection years from 2020}

\item{dynamicincdebt}{dynamic input of increase in trailing cost of debt}
}
\description{
Trailing average portfolio return on debt for each year in model
}
\details{
rate of debt changes with year starting from the first value (debt start) and reducing by debt reduce
rate of debt is capped at the fifth year and then made into a dataframe with years as column headings.

Prerequisites that need to be defined in ptrm model (i.e. run above code):
year, debtstart, debtfinal, debtreduce, startyearend, projyearend
}
\keyword{debt}
\keyword{debt,}
\keyword{on}
\keyword{rate,}
\keyword{return}
